move classes to own namespaces
