+++ Task_02/Printers/WinRatePrinter.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections.Generic;
+using Task_02.Statistics;
+
+namespace Task_02.Printers
+{
+    internal abstract class WinRatePrinter : IPrinter
+    {
+        private string _winRateFormat = " {0} (винрейт {1:0.00})";
+
+        public abstract void Print(IEnumerable<HeroScoreInfo> heroes);
+
+        protected string GetWinRateString(HeroScoreInfo heroScoreInfo)
+        {
+            return string.Format(_winRateFormat, heroScoreInfo.HeroName, heroScoreInfo.GetWinRate());
+        }
+    }
+}
