+++ Task_02/Program.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using Task_02.Printers;
+using Task_02.Statistics;
 
 namespace Task_02
 {
