+++ Task_02/Statistics/HeroUsagesStatisticMostUnfavoriteProvider.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+using Task_02.Printers;
+
+namespace Task_02.Statistics
+{
+    internal class HeroUsagesStatisticMostUnfavoriteProvider : AbstractStatisticProvider
+    {
+        public HeroUsagesStatisticMostUnfavoriteProvider(List<HeroScoreInfo> infos) : base(infos, new UnfavoritePrinter()) { }
+
+        internal override void Show()
+        {
+            Printer.Print(Calculator.FindMostUnfavoriteHeros(Infos));
+        }
+    }
+}
