+++ Task_02/Statistics/HeroUsagesStatisticProvider.cs
@@ -1,8 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
+using Task_02.Printers;
+using Task_02.Statistics;
 
 namespace Task_02
 {
